Move DB utilities from `utils` to `db`
The problem with `utils` modules is that they very quickly become
disorganised. Let's move those utilities that relate to databases to
their own module, which we'll call `db`.
