+++ tasks/tasks/get_project_definitions.py
@@ -3,7 +3,7 @@ import collections
 import pipeline
 import sqlalchemy
 
-from .. import DATA_DIR, INDEX_DATE, io, utils
+from .. import DATA_DIR, INDEX_DATE, db, io, utils
 
 
 Record = collections.namedtuple("Record", ["repo", "sha", "project_definition"])
@@ -43,8 +43,8 @@ def write_pickle(records, project_definitions_dir):
 
 def main():  # pragma: no cover
     # This is hard to test without a Job Server DB, so we exclude it from coverage.
-    engine = utils.get_engine(utils.Database.JOBSERVER)
-    metadata = utils.get_metadata(engine)
+    engine = db.get_engine(db.Database.JOBSERVER)
+    metadata = db.get_metadata(engine)
 
     rows = extract(engine, metadata)
     records = (get_record(row) for row in rows)
