+++ tests/tasks/test_utils.py
@@ -1,30 +1,4 @@
-import pytest
-
 from tasks import io, utils
-
-
-@pytest.mark.parametrize(
-    "database, environment_variable",
-    [
-        (utils.Database.JOBSERVER, "JOBSERVER_DATABASE_URL"),
-        (utils.Database.OPENCODELISTS, "OPENCODELISTS_DATABASE_URL"),
-    ],
-)
-def test_get_engine(database, environment_variable, monkeypatch):
-    monkeypatch.setenv(environment_variable, "sqlite+pysqlite:///:memory:")
-    engine = utils.get_engine(database)
-    assert str(engine.url) == "sqlite+pysqlite:///:memory:"
-
-
-def test_get_engine_when_unknown_database():
-    with pytest.raises(TypeError, match="Cannot get engine for unknown database: foo"):
-        utils.get_engine("foo")
-
-
-def test_get_metadata(monkeypatch):
-    monkeypatch.setenv("JOBSERVER_DATABASE_URL", "sqlite+pysqlite:///:memory:")
-    metadata = utils.get_metadata(utils.get_engine(utils.Database.JOBSERVER))
-    assert metadata.tables == {}
 
 
 def test_get_repo():
